+++ ptai-teamcity-plugin/ptai-teamcity-plugin-common/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/teamcity/Labels.java
@@ -140,14 +140,29 @@ public class Labels {
     public static final String REPORTING_RAWDATA_FILE = Resources.i18n_ast_settings_mode_synchronous_subjob_export_rawjson_file_label();
 
     /**
+     * See {@link Params#REPORTING_RAWDATA_FILTER}
+     */
+    public static final String REPORTING_RAWDATA_FILTER = Resources.i18n_ast_settings_mode_synchronous_subjob_export_rawjson_filter_label();
+
+    /**
      * See {@link Params#REPORTING_SARIF_FILE}
      */
     public static final String REPORTING_SARIF_FILE = Resources.i18n_ast_settings_mode_synchronous_subjob_export_sarif_file_label();
 
     /**
+     * See {@link Params#REPORTING_SARIF_FILTER}
+     */
+    public static final String REPORTING_SARIF_FILTER = Resources.i18n_ast_settings_mode_synchronous_subjob_export_sarif_filter_label();
+
+    /**
      * See {@link Params#REPORTING_SONARGIIF_FILE}
      */
     public static final String REPORTING_SONARGIIF_FILE = Resources.i18n_ast_settings_mode_synchronous_subjob_export_sonargiif_file_label();
+
+    /**
+     * See {@link Params#REPORTING_SONARGIIF_FILTER}
+     */
+    public static final String REPORTING_SONARGIIF_FILTER = Resources.i18n_ast_settings_mode_synchronous_subjob_export_sonargiif_filter_label();
 
     /**
      * See {@link Params#REPORTING_JSON_SETTINGS}