+++ ptai-teamcity-plugin/ptai-teamcity-plugin-common/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/teamcity/Params.java
@@ -184,16 +184,34 @@ public class Params {
     public static final String REPORTING_RAWDATA_FILE = PARAM("ReportingRawDataFile");
 
     /**
+     * If {@link Params#REPORTING_RAWDATA} is on then this field may
+     * contain JSON filter to define vulnerabilities to be included in the generated raw data export file
+     */
+    public static final String REPORTING_RAWDATA_FILTER = PARAM("ReportingRawDataFilter");
+
+    /**
      * If {@link Params#REPORTING_SARIF} is on then this field is to
      * contain name of the file where raw SARIF JSON data will be saved to
      */
     public static final String REPORTING_SARIF_FILE = PARAM("ReportingSarifFile");
 
     /**
+     * If {@link Params#REPORTING_SARIF} is on then this field may
+     * contain JSON filter to define vulnerabilities to be included in the generated SARIF export file
+     */
+    public static final String REPORTING_SARIF_FILTER = PARAM("ReportingSarifFilter");
+
+    /**
      * If {@link Params#REPORTING_SONARGIIF} is on then this field is to
      * contain name of the file where SonarQube GIIF JSON data will be saved to
      */
     public static final String REPORTING_SONARGIIF_FILE = PARAM("ReportingSonarGiifFile");
+
+    /**
+     * If {@link Params#REPORTING_SONARGIIF} is on then this field may
+     * contain JSON filter to define vulnerabilities to be included in the generated SonarQube GIIF export file
+     */
+    public static final String REPORTING_SONARGIIF_FILTER = PARAM("ReportingSonarGiifFilter");
 
     /**
      * If {@link Params#REPORTING_JSON} is on then this field is to